Breakup monolithic imports
Move commonly used code out of monolithic imports to reduce the amount of code imported by the `.cache.ts` files.

Add a new `MUST_NOT_BE_IMPORTED_BY_DATA_CACHE` token that can be used to figure out why the cache wants to regenerate for a given file. Add this to device_pool.ts, as this really shouldn't be depended on for the cache.
