+++ src/webgpu/shader/execution/expression/binary/i32_arithmetic.cache.ts
@@ -1,11 +1,11 @@
 import { kValue } from '../../../../util/constants.js';
 import { sparseI32Range, vectorI32Range } from '../../../../util/math.js';
-import { makeCaseCache } from '../case_cache.js';
 import {
   generateBinaryToI32Cases,
   generateI32VectorBinaryToVectorCases,
   generateVectorI32BinaryToVectorCases,
-} from '../expression.js';
+} from '../case.js';
+import { makeCaseCache } from '../case_cache.js';
 
 function i32_add(x: number, y: number): number | undefined {
   return x + y;
