+++ src/webgpu/shader/execution/expression/binary/u32_arithmetic.cache.ts
@@ -1,10 +1,10 @@
 import { sparseU32Range, vectorU32Range } from '../../../../util/math.js';
-import { makeCaseCache } from '../case_cache.js';
 import {
   generateBinaryToU32Cases,
   generateU32VectorBinaryToVectorCases,
   generateVectorU32BinaryToVectorCases,
-} from '../expression.js';
+} from '../case.js';
+import { makeCaseCache } from '../case_cache.js';
 
 function u32_add(x: number, y: number): number | undefined {
   return x + y;
