+++ src/webgpu/shader/execution/expression/call/builtin/max.spec.ts
@@ -27,7 +27,8 @@ import {
   i32,
   u32,
 } from '../../../../../util/conversion.js';
-import { Case, allInputSources, onlyConstInputSource, run } from '../../expression.js';
+import { Case } from '../../case.js';
+import { allInputSources, onlyConstInputSource, run } from '../../expression.js';
 
 import { abstractBuiltin, builtin } from './builtin.js';
 import { d } from './max.cache.js';
