+++ src/webgpu/shader/execution/expression/call/builtin/modf.cache.ts
@@ -1,7 +1,7 @@
 import { toVector } from '../../../../../util/conversion.js';
 import { FP, FPKind } from '../../../../../util/floating_point.js';
+import { Case } from '../../case.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { Case } from '../../expression.js';
 
 /** @returns a fract Case for a scalar vector input */
 function makeScalarCaseFract(kind: FPKind, n: number): Case {
