+++ src/webgpu/shader/execution/expression/call/builtin/pack2x16snorm.spec.ts
@@ -18,7 +18,8 @@ import {
   vec2,
 } from '../../../../../util/conversion.js';
 import { quantizeToF32, vectorF32Range } from '../../../../../util/math.js';
-import { allInputSources, Case, run } from '../../expression.js';
+import { Case } from '../../case.js';
+import { allInputSources, run } from '../../expression.js';
 
 import { builtin } from './builtin.js';
