+++ src/webgpu/shader/execution/expression/call/builtin/refract.cache.ts
@@ -1,8 +1,9 @@
 import { ROArrayArray } from '../../../../../../common/util/types.js';
 import { toVector } from '../../../../../util/conversion.js';
 import { FP, FPKind } from '../../../../../util/floating_point.js';
+import { Case } from '../../case.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { Case, IntervalFilter } from '../../expression.js';
+import { IntervalFilter } from '../../interval_filter.js';
 
 // Using a bespoke implementation of make*Case and generate*Cases here
 // since refract is the only builtin with the API signature
