+++ src/webgpu/shader/execution/expression/call/builtin/step.cache.ts
@@ -1,7 +1,7 @@
 import { anyOf } from '../../../../../util/compare.js';
 import { FP } from '../../../../../util/floating_point.js';
+import { Case } from '../../case.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { Case } from '../../expression.js';
 
 // stepInterval's return value can't always be interpreted as a single acceptance
 // interval, valid result may be 0.0 or 1.0 or both of them, but will never be a
