+++ src/webgpu/util/device_pool.ts
@@ -9,6 +9,9 @@ import {
 } from '../../common/util/util.js';
 import { getDefaultLimits, kLimits } from '../capability_info.js';
 
+// MUST_NOT_BE_IMPORTED_BY_DATA_CACHE
+// This file should not be transitively imported by .cache.ts files
+
 export interface DeviceProvider {
   readonly device: GPUDevice;
   expectDeviceLost(reason: GPUDeviceLostReason): void;
